+++ lib/karafka/base_controller.rb
@@ -3,10 +3,10 @@
 # Karafka module namespace
 module Karafka
   # Base controller from which all Karafka controllers should inherit
-  # Similar to Rails controllers we can define before_enqueue callbacks
+  # Similar to Rails controllers we can define after_received callbacks
   # that will be executed
   #
-  # Note that if before_enqueue return false, the chain will be stopped and
+  # Note that if after_received return false, the chain will be stopped and
   #   the perform method won't be executed in sidekiq (won't peform_async it)
   #
   # @example Create simple controller
@@ -16,9 +16,9 @@ module Karafka
   #     end
   #   end
   #
-  # @example Create a controller with a block before_enqueue
+  # @example Create a controller with a block after_received
   #   class ExampleController < Karafka::BaseController
-  #     before_enqueue do
+  #     after_received do
   #       # Here we should have some checking logic
   #       # If false is returned, won't schedule a perform action
   #     end
@@ -28,9 +28,9 @@ module Karafka
   #     end
   #   end
   #
-  # @example Create a controller with a method before_enqueue
+  # @example Create a controller with a method after_received
   #   class ExampleController < Karafka::BaseController
-  #     before_enqueue :before_method
+  #     after_received :after_received_method
   #
   #     def perform
   #       # some logic here
@@ -38,52 +38,55 @@ module Karafka
   #
   #     private
   #
-  #     def before_method
+  #     def after_received_method
   #       # Here we should have some checking logic
   #       # If false is returned, won't schedule a perform action
-  #     end
-  #   end
-  #
-  # @example Create a controller with an after_failure action
-  #   class ExampleController < Karafka::BaseController
-  #     def perform
-  #       # some logic here
-  #     end
-  #
-  #     def after_failure
-  #       # action taken in case perform fails
   #     end
   #   end
   class BaseController
     extend ActiveSupport::DescendantsTracker
     include ActiveSupport::Callbacks
 
-    # The schedule method is wrapped with a set of callbacks
+    # The call method is wrapped with a set of callbacks
     # We won't run perform at the backend if any of the callbacks
     # returns false
     # @see http://api.rubyonrails.org/classes/ActiveSupport/Callbacks/ClassMethods.html#method-i-get_callbacks
-    define_callbacks :schedule
+    define_callbacks :after_received
 
-    # Each controller instance is always bind to a single topic. We don't place it on a class
-    # level because some programmers use same controller for multiple topics
-    attr_accessor :topic
     attr_accessor :params_batch
 
     class << self
+      attr_reader :topic
+
+      # Assigns a topic to a controller and build up proper controller functionalities, so it can
+      #   cooperate with the topic settings
+      # @param topic [Karafka::Routing::Topic]
+      # @return [Karafka::Routing::Topic] assigned topic
+      def topic=(topic)
+        @topic = topic
+        Controllers::Includer.call(self)
+        @topic
+      end
+
       # Creates a callback that will be executed before scheduling to Sidekiq
       # @param method_name [Symbol, String] method name or nil if we plan to provide a block
       # @yield A block with a code that should be executed before scheduling
       # @note If value returned is false, will chalt the chain and not schedlue to Sidekiq
-      # @example Define a block before_enqueue callback
-      #   before_enqueue do
+      # @example Define a block after_received callback
+      #   after_received do
       #     # logic here
       #   end
       #
-      # @example Define a class name before_enqueue callback
-      #   before_enqueue :method_name
-      def before_enqueue(method_name = nil, &block)
-        set_callback :schedule, :before, method_name ? method_name : block
+      # @example Define a class name after_received callback
+      #   after_received :method_name
+      def after_received(method_name = nil, &block)
+        set_callback :after_received, :before, method_name ? method_name : block
       end
+    end
+
+    # @return [Karafka::Routing::Topic] topic to which a given controller is subscribed
+    def topic
+      self.class.topic
     end
 
     # Creates lazy loaded params batch object
@@ -95,33 +98,12 @@ module Karafka
       @params_batch = Karafka::Params::ParamsBatch.new(messages, topic.parser)
     end
 
-    # @return [Karafka::Params::Params] params instance for non batch processed controllers
-    # @raise [Karafka::Errors::ParamsMethodUnavailable] raised when we try to use params
-    #   method in a batch_processed controller
-    def params
-      raise Karafka::Errors::ParamsMethodUnavailable if topic.batch_processing
-      params_batch.first
-    end
-
     # Executes the default controller flow, runs callbacks and if not halted
-    # will schedule a call task in sidekiq
-    def schedule
-      run_callbacks :schedule do
-        case topic.processing_backend
-        when :inline then
-          call_inline
-        when :sidekiq then
-          call_async
-        else
-          raise Errors::InvalidProcessingBackend, topic.processing_backend
-        end
-      end
-    end
-
-    # @note We want to leave the #perform method as a public API, but just in case we will do some
-    #   pre or post processing we use call method instead of directly executing #perform
+    # will call process method of a proper backend
     def call
-      perform
+      run_callbacks :after_received do
+        process
+      end
     end
 
     private
@@ -131,44 +113,6 @@ module Karafka
     #   someone forgets about it or makes on with typo
     def perform
       raise NotImplementedError, 'Implement this in a subclass'
-    end
-
-    # @return [Karafka::BaseResponder] responder instance if defined
-    # @return [nil] nil if no responder for this controller
-    def responder
-      @responder ||= topic.responder&.new(topic.parser)
-    end
-
-    # Responds with given data using given responder. This allows us to have a similar way of
-    # defining flows like synchronous protocols
-    # @param data Anything we want to pass to responder based on which we want to trigger further
-    #   Kafka responding
-    # @raise [Karafka::Errors::ResponderMissing] raised when we don't have a responder defined,
-    #   but we still try to use this method
-    def respond_with(*data)
-      raise(Errors::ResponderMissing, self.class) unless responder
-      Karafka.monitor.notice(self.class, data: data)
-      responder.call(*data)
-    end
-
-    # Executes perform code immediately (without enqueuing)
-    # @note Despite the fact, that workers won't be used, we still initialize all the
-    #   classes and other framework elements
-    def call_inline
-      Karafka.monitor.notice(self.class, params_batch)
-      call
-    end
-
-    # Enqueues the execution of perform method into a worker.
-    # @note Each worker needs to have a class #perform_async method that will allow us to pass
-    #   parameters into it. We always pass topic as a first argument and this request params_batch
-    #   as a second one (we pass topic to be able to build back the controller in the worker)
-    def call_async
-      Karafka.monitor.notice(self.class, params_batch)
-      topic.worker.perform_async(
-        topic.id,
-        topic.interchanger.load(params_batch.to_a)
-      )
     end
   end
 end