+++ lib/karafka/connection/messages_processor.rb
@@ -18,12 +18,16 @@ module Karafka
           # first one and it will be valid for all the messages
           # We map from incoming topic name, as it might be namespaced, etc.
           # @see topic_mapper internal docs
-          mapped_topic = Karafka::App.config.topic_mapper.incoming(kafka_messages[0].topic)
-          # @note We search based on the topic id - that is a combination of group id and
-          # topic name
-          controller = Karafka::Routing::Router.build("#{group_id}_#{mapped_topic}")
-          handler = controller.topic.batch_processing ? :process_batch : :process_each
+          mapped_topic_name = Karafka::App.config.topic_mapper.incoming(kafka_messages[0].topic)
+          topic = Routing::Router.find("#{group_id}_#{mapped_topic_name}")
 
+          # Depending on a case (persisted or not) we might use new controller instance per each
+          # batch, or use the same instance for all of them (for implementing buffering, etc)
+          controller = Persistence.fetch(topic, kafka_messages[0].partition, :controller) do
+            topic.controller.new
+          end
+
+          handler = topic.batch_processing ? :process_batch : :process_each
           send(handler, controller, kafka_messages)
         end
 
@@ -35,7 +39,7 @@ module Karafka
         def process_batch(controller, kafka_messages)
           controller.params_batch = kafka_messages
           Karafka.monitor.notice(self, kafka_messages)
-          controller.schedule
+          controller.call
         end
 
         # Processes messages one by one (like with std http requests)