+++ lib/karafka/setup/config.rb
@@ -42,6 +42,13 @@ module Karafka
       # #params_batch will contain params received from Kafka (may be more than 1) so we can
       # process them in batches
       setting :batch_processing, false
+      # Should we operate in a single controller instance across multiple batches of messages,
+      # from the same partition or should we build a new instance for each incoming batch.
+      # Disabling that can be useful when you want to build a new controller instance for each
+      # incoming batch. It's disabled by default, not to create more objects that needed on
+      # each batch
+      # @note It won't have any effect on Sidekiq backend
+      setting :persistent, true
       # Connection pool options are used for producer (Waterdrop)
       # They are configured automatically based on Sidekiq concurrency and number of consumers
       # The bigger one is selected as we need to be able to send messages from both places